modified post api
