+++ config/settings/base.py
@@ -51,6 +51,7 @@ CUSTOM_APPS = [
     "corsheaders",
     "rest_framework_simplejwt",
     "rest_framework_simplejwt.token_blacklist",
+    "images.apps.ImagesConfig",
 ]
