+++ posts/urls.py
@@ -1,18 +1,11 @@
 from django.urls import path
-from .views import (
-    TripPostCreateView,
-    TripPostDetailView,
-    TripPostUpdateView,
-    TripPostDeleteView,
-    TripPostListView,
-    UserPostListView
-)
+from .views import TripPostView, PostDetailView, PostListView, Thumbnail
+from rest_framework_simplejwt.views import TokenObtainPairView
 
 urlpatterns = [
-    path("", TripPostCreateView.as_view(), name="create_post"),
-    path("<int:post_id>/", TripPostDetailView.as_view(), name="post_detail"),
-    path("<int:post_id>/modify/", TripPostUpdateView.as_view(), name="post_update"),
-    path("<int:post_id>/delete/", TripPostDeleteView.as_view(), name="post_delete"),
-    path("all/", TripPostListView.as_view(), name="all_post_list"),
-    path("<int:user_id>/post/", UserPostListView.as_view(), name="user_post_list"),
+    path('', TripPostView.as_view(), name='post-list'),
+    path('login/', TokenObtainPairView.as_view(), name='post-login'),
+    path('<int:post_id>/', PostDetailView.as_view(), name='post-detail'),
+    path('list/<int:user_id>/', PostListView.as_view(), name='post-list'),
+    path('thumbnails/', Thumbnail.as_view(), name='post-thumbnails'),
 ]
